AudiobookLibrarySource: Fix crash when adding tracks (bgo#662906)
We need to cast the TrackMediaAttributes enum value as an int, otherwise
it is handled as string and the query fails.

+++ src/Extensions/Banshee.Audiobook/Banshee.Audiobook/AudiobookLibrarySource.cs
@@ -124,7 +124,7 @@ namespace Banshee.Audiobook
 
                     ServiceManager.DbConnection.Execute (
                         "UPDATE CoreTracks SET Attributes = Attributes | ? WHERE PrimarySourceID = ?",
-                        TrackMediaAttributes.AudioBook, this.DbId);
+                        (int)TrackMediaAttributes.AudioBook, this.DbId);
                 }
             };
